+++ source/pipes/math/date-time-latest.test.ts
@@ -12,6 +12,10 @@ describe("DateTimeLatestPipe", () => {
 
   describe("#transform", () => {
 
+    it("transforms an empty list of DateTimes into undefined", () => {
+      expect(pipe.transform([])).toBeUndefined();
+    });
+
     it("transforms a list of DateTimes into the maximum of that list", () => {
       const result = pipe.transform([
         DateTime.fromISO("2006-01-02T15:04:05-07:00"),