+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project are documented here. The format is based on 
 ## Unreleased
 
 -   Added Duration -> arbitrary format pipe.
+-   Added longest/shortest Duration pipes.
 
 ## 0.4.0 (2018-05-06)
 
@@ -21,7 +22,7 @@ All notable changes to this project are documented here. The format is based on 
 -   Aligned to Luxon 0.2.8.
 -   Added milliseconds -> DateTime pipe.
 -   Added SQL <-> DateTime pipes.
--   Added earliest/latest pipes.
+-   Added earliest/latest DateTime pipes.
 
 ## 0.1.0 (2017-10-23)